"LIFESTREAM: BLOOD DONATION WEBSITE"
The problem statement for the LIFESTREAM Blood Donation website is about creating a website where people who want to donate blood can easily connect with those who need blood. The goal is to make it simple for donors to find recipients who need their blood type, and for recipients to find donors. The website should also make sure donors are safe and their information is kept private. Additionally, it should help people understand why donating blood is important. Overall, the website's purpose is to make it easy for people to donate blood and save lives. Our aim is to create a user-friendly online platform that connects blood donors with recipients in need of blood transfusions.